chore: remove admn permissions and check in service

+++ src/routes/league/admin/admin.service.ts
@@ -14,6 +14,14 @@ export class AdminService {
         @InjectRepository(User) private userDb: Repository<User>
     ) {}
 
+    private async checkHeadAdmin(leagueId: number, discordId: string) {
+        const data = await this.leagueAdminDb
+            .createQueryBuilder('user')
+            .where('user.discord_id = :discordId AND user.league_id = :leagueId', { discordId, leagueId })
+            .getOne();
+        if (!data.headAdmin) throw new HttpException("You are't allowed to perform this action.", HttpStatus.UNAUTHORIZED);
+    }
+
     public async admins(leagueId: number) {
         const query = `SELECT t1.id, t1.discord_id AS "discordId", t1.league_id AS "leagueId", t1.permissions, t1.head_admin AS "headAdmin", t1.added_at AS "addedAt", 
 			t2.user_name AS "username", t2.discriminator, t2.avatar 
@@ -22,7 +30,8 @@ export class AdminService {
         return await this.db.query(query, [leagueId]);
     }
 
-    public async addAdmin(discordId: string, leagueId: number, permissions: number) {
+    public async addAdmin(userDiscordId: string, discordId: string, leagueId: number, permissions: number) {
+        await this.checkHeadAdmin(leagueId, userDiscordId);
         const check = await this.userDb.query('SELECT EXISTS(SELECT 1 FROM users WHERE discord_id = $1)', [discordId]);
 
         if (!check) throw new HttpException('User is not yet registered on this site.', HttpStatus.NOT_FOUND);
@@ -43,7 +52,8 @@ export class AdminService {
         }
     }
 
-    public async updateAdminPermissions(leagueId: number, adminId: number, permissions: number) {
+    public async updateAdminPermissions(userDiscordId: string, leagueId: number, adminId: number, permissions: number) {
+        await this.checkHeadAdmin(leagueId, userDiscordId);
         await this.leagueAdminDb
             .createQueryBuilder()
             .update()
@@ -52,7 +62,8 @@ export class AdminService {
             .execute();
     }
 
-    public async removeAdmin(adminId: number, leagueId: number) {
+    public async removeAdmin(userDiscordId: string, adminId: number, leagueId: number) {
+        await this.checkHeadAdmin(leagueId, userDiscordId);
         await this.leagueAdminDb.query('DELETE FROM league_admin WHERE id = $1 AND league_id = $2', [adminId, leagueId]);
     }
 }
